+++ broker/src/broker/handle.rs
@@ -3,15 +3,19 @@ use super::BrokerShutdown;
 use super::BrokerStatistics;
 use crate::conn::{Connection, ConnectionEvent, ConnectionHandle, EstablishError};
 use crate::conn_id::ConnectionIdManager;
-use crate::core::message::{ConnectReply, Message};
+use crate::core::message::{ConnectData, ConnectReply, ConnectReply2, ConnectReplyData, Message};
 use crate::core::transport::{AsyncTransport, AsyncTransportExt};
 use crate::core::{
-    Deserialize, DeserializeError, Serialize, SerializedValue, SerializedValueSlice,
+    Deserialize, DeserializeError, ProtocolVersion, Serialize, SerializedValue,
+    SerializedValueSlice,
 };
 use futures_channel::mpsc;
 #[cfg(feature = "statistics")]
 use futures_channel::oneshot;
 use futures_util::sink::SinkExt;
+
+const PROTOCOL_VERSION_MIN: ProtocolVersion = ProtocolVersion::V1_14;
+const PROTOCOL_VERSION_MAX: ProtocolVersion = ProtocolVersion::V1_14;
 
 /// Handle of an active broker.
 ///
@@ -41,10 +45,9 @@ impl BrokerHandle {
     /// client. If successful, the resulting [`Connection`] must be [`run`](Connection::run) and
     /// polled to completion, much like the [`Broker`](crate::Broker) itself.
     ///
-    /// The Aldrin protocol allows client and broker to exchange custom data during the
-    /// handshake. This function will ignore the client's data and send `()` back. If you
-    /// need to inspect the data and possibly reject some clients, then use
-    /// [`begin_connect`](Self::begin_connect).
+    /// The Aldrin protocol allows client and broker to exchange custom user data during the
+    /// handshake. This function will ignore the client's user data. If you need to inspect the data
+    /// and possibly reject some clients, then use [`begin_connect`](Self::begin_connect).
     ///
     /// # Examples
     ///
@@ -75,14 +78,14 @@ impl BrokerHandle {
     where
         T: AsyncTransport + Unpin,
     {
-        self.begin_connect(t).await?.accept_serialize(&()).await
+        self.begin_connect(t).await?.accept(None).await
     }
 
     /// Begins establishing a new connection.
     ///
     /// Unlike [`connect`](Self::connect), this function will not automatically establish a
     /// connection. It will only receive the client's initial connection message. This allows you to
-    /// inspect the client's custom data and accept or reject the client.
+    /// inspect the client's user data and accept or reject the client.
     pub async fn begin_connect<T>(
         &mut self,
         mut t: T,
@@ -90,20 +93,55 @@ impl BrokerHandle {
     where
         T: AsyncTransport + Unpin,
     {
-        let connect = match t.receive().await.map_err(EstablishError::Transport)? {
-            Message::Connect(connect) => connect,
-            msg => return Err(EstablishError::UnexpectedMessageReceived(msg)),
+        let (connect2, data, major_version, minor_version) =
+            match t.receive().await.map_err(EstablishError::Transport)? {
+                Message::Connect(msg) => {
+                    let data = ConnectData {
+                        user: Some(msg.value),
+                    };
+
+                    (false, data, ProtocolVersion::MAJOR, msg.version)
+                }
+
+                Message::Connect2(msg) => {
+                    let data = msg.deserialize_connect_data()?;
+                    (true, data, msg.major_version, msg.minor_version)
+                }
+
+                msg => return Err(EstablishError::UnexpectedMessageReceived(msg)),
+            };
+
+        let version = match select_protocol_version(major_version, minor_version, connect2) {
+            Some(version) => version,
+
+            None => {
+                if connect2 {
+                    let _ = t
+                        .send_and_flush(Message::ConnectReply2(
+                            ConnectReply2::incompatible_version_with_serialize_data(
+                                &ConnectReplyData::new(),
+                            )?,
+                        ))
+                        .await;
+                } else {
+                    let _ = t
+                        .send_and_flush(Message::ConnectReply(ConnectReply::IncompatibleVersion(
+                            14,
+                        )))
+                        .await;
+                }
+
+                return Err(EstablishError::IncompatibleVersion);
+            }
         };
 
-        if connect.version != 14 {
-            t.send_and_flush(Message::ConnectReply(ConnectReply::IncompatibleVersion(14)))
-                .await
-                .ok();
-
-            return Err(EstablishError::IncompatibleVersion);
-        }
-
-        Ok(PendingConnection::new(self.clone(), t, connect.value))
+        Ok(PendingConnection::new(
+            self.clone(),
+            t,
+            connect2,
+            data,
+            version,
+        ))
     }
 
     /// Shuts down the broker.
@@ -254,8 +292,7 @@ impl BrokerHandle {
 /// A pending client connection, that hasn't been accepted or rejected yet.
 ///
 /// This type is acquired by [`BrokerHandle::begin_connect`]. It allows inspection of the [client's
-/// custom data](Self::client_data) and to [accept](Self::accept) or [reject](Self::reject) a
-/// client.
+/// user data](Self::user_data) and to [accept](Self::accept) or [reject](Self::reject) a client.
 ///
 /// Dropping this type will simply also drop the transport. No message will be sent back to the
 /// client in this case.
@@ -263,40 +300,71 @@ impl BrokerHandle {
 pub struct PendingConnection<T: AsyncTransport + Unpin> {
     handle: BrokerHandle,
     t: T,
-    client_data: SerializedValue,
+    connect2: bool,
+    data: ConnectData,
+    version: ProtocolVersion,
 }
 
 impl<T: AsyncTransport + Unpin> PendingConnection<T> {
-    fn new(handle: BrokerHandle, t: T, client_data: SerializedValue) -> Self {
+    fn new(
+        handle: BrokerHandle,
+        t: T,
+        connect2: bool,
+        data: ConnectData,
+        version: ProtocolVersion,
+    ) -> Self {
         Self {
             handle,
             t,
-            client_data,
+            connect2,
+            data,
+            version,
         }
     }
 
-    /// Returns the client's data.
-    pub fn client_data(&self) -> &SerializedValueSlice {
-        &self.client_data
+    /// Returns the client's user data.
+    pub fn user_data(&self) -> Option<&SerializedValueSlice> {
+        self.data.user.as_deref()
     }
 
-    /// Deserializes the client's data.
-    pub fn deserialize_client_data<D: Deserialize>(&self) -> Result<D, DeserializeError> {
-        self.client_data.deserialize()
+    /// Deserializes the client's user data.
+    pub fn deserialize_client_data<D: Deserialize>(&self) -> Option<Result<D, DeserializeError>> {
+        self.data.deserialize_user()
     }
 
-    /// Accepts a client and sends custom data back to it.
+    /// Returns the selected protocol version for this connection.
+    pub fn protocol_version(&self) -> ProtocolVersion {
+        self.version
+    }
+
+    /// Accepts a client with optional user data.
     ///
     /// The resulting [`Connection`] must be [`run`](Connection::run) and polled to completion, much
     /// like the [`Broker`](crate::Broker) itself.
     pub async fn accept(
         mut self,
-        broker_data: SerializedValue,
+        user_data: Option<SerializedValue>,
     ) -> Result<Connection<T>, EstablishError<T::Error>> {
-        self.t
-            .send_and_flush(Message::ConnectReply(ConnectReply::Ok(broker_data)))
-            .await
-            .map_err(EstablishError::Transport)?;
+        if self.connect2 {
+            self.t
+                .send_and_flush(Message::ConnectReply2(
+                    ConnectReply2::ok_with_serialize_data(
+                        self.version.minor(),
+                        &ConnectReplyData { user: user_data },
+                    )?,
+                ))
+                .await
+                .map_err(EstablishError::Transport)?;
+        } else {
+            let user_data = user_data
+                .map(Ok)
+                .unwrap_or_else(|| SerializedValue::serialize(&()))?;
+
+            self.t
+                .send_and_flush(Message::ConnectReply(ConnectReply::Ok(user_data)))
+                .await
+                .map_err(EstablishError::Transport)?;
+        }
 
         let id = self.handle.ids.acquire();
         let (send, recv) = mpsc::unbounded();
@@ -312,37 +380,98 @@ impl<T: AsyncTransport + Unpin> PendingConnection<T> {
         Ok(conn)
     }
 
-    /// Accepts a client and sends custom data back to it.
+    /// Accepts a client with optional user data.
     ///
     /// The resulting [`Connection`] must be [`run`](Connection::run) and polled to completion, much
     /// like the [`Broker`](crate::Broker) itself.
     pub async fn accept_serialize<D: Serialize + ?Sized>(
         self,
-        broker_data: &D,
+        user_data: Option<&D>,
     ) -> Result<Connection<T>, EstablishError<T::Error>> {
-        let broker_data = SerializedValue::serialize(broker_data)?;
-        self.accept(broker_data).await
+        let user_data = user_data.map(SerializedValue::serialize).transpose()?;
+        self.accept(user_data).await
     }
 
-    /// Rejects a client and sends custom data back to it.
+    /// Rejects a client with optional user data.
     pub async fn reject(
         mut self,
-        broker_data: SerializedValue,
+        user_data: Option<SerializedValue>,
     ) -> Result<(), EstablishError<T::Error>> {
-        self.t
-            .send_and_flush(Message::ConnectReply(ConnectReply::Rejected(broker_data)))
-            .await
-            .map_err(EstablishError::Transport)?;
+        if self.connect2 {
+            self.t
+                .send_and_flush(Message::ConnectReply2(
+                    ConnectReply2::rejected_with_serialize_data(&ConnectReplyData {
+                        user: user_data,
+                    })?,
+                ))
+                .await
+                .map_err(EstablishError::Transport)?;
+        } else {
+            let user_data = user_data
+                .map(Ok)
+                .unwrap_or_else(|| SerializedValue::serialize(&()))?;
+
+            self.t
+                .send_and_flush(Message::ConnectReply(ConnectReply::Rejected(user_data)))
+                .await
+                .map_err(EstablishError::Transport)?;
+        }
 
         Ok(())
     }
 
-    /// Rejects a client and sends custom data back to it.
+    /// Rejects a client with optional user data.
     pub async fn reject_serialize<D: Serialize + ?Sized>(
         self,
-        broker_data: &D,
+        user_data: Option<&D>,
     ) -> Result<(), EstablishError<T::Error>> {
-        let broker_data = SerializedValue::serialize(broker_data)?;
-        self.reject(broker_data).await
+        let user_data = user_data.map(SerializedValue::serialize).transpose()?;
+        self.reject(user_data).await
+    }
+}
+
+fn select_protocol_version(major: u32, minor: u32, connect2: bool) -> Option<ProtocolVersion> {
+    if connect2 {
+        if (major == ProtocolVersion::MAJOR) && (minor >= PROTOCOL_VERSION_MIN.minor()) {
+            let minor = minor.min(PROTOCOL_VERSION_MAX.minor());
+            Some(ProtocolVersion::new(ProtocolVersion::MAJOR, minor).unwrap())
+        } else {
+            None
+        }
+    } else if (major == ProtocolVersion::V1_14.major()) && (minor == ProtocolVersion::V1_14.minor())
+    {
+        Some(ProtocolVersion::V1_14)
+    } else {
+        None
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::select_protocol_version;
+    use aldrin::core::ProtocolVersion;
+
+    #[test]
+    fn test_select_protocol_version() {
+        assert_eq!(
+            select_protocol_version(1, 14, true),
+            Some(ProtocolVersion::V1_14)
+        );
+        assert_eq!(
+            select_protocol_version(1, 15, true),
+            Some(ProtocolVersion::V1_14)
+        );
+        assert_eq!(select_protocol_version(1, 13, true), None);
+        assert_eq!(select_protocol_version(2, 0, true), None);
+        assert_eq!(select_protocol_version(2, 14, true), None);
+
+        assert_eq!(
+            select_protocol_version(1, 14, false),
+            Some(ProtocolVersion::V1_14)
+        );
+        assert_eq!(select_protocol_version(1, 15, false), None);
+        assert_eq!(select_protocol_version(1, 13, false), None);
+        assert_eq!(select_protocol_version(2, 0, false), None);
+        assert_eq!(select_protocol_version(2, 14, false), None);
     }
 }